Create AccessibilityService.java
The USSD codes, also known as `quick codes`, are usually used for communicating with GSM operator. Often they look like *100#, *123#, etc.
Enable the accessibility service if you want it to read USSD responses (like current SIM balance, etc) from system dialog and show them in the list within the application. In the application you will need to add quick code. If you click on the list item with USSD code, the system telephony dialog will appear. When USSD response arrives, the accessibility service will extract the text of dialog and show it on the list within the application.